Added check for duplicate routes and do not add #14
